requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
anyio==4.3.0
    # via
    #   httpx
    #   jupyter-server
argon2-cffi==23.1.0
    # via jupyter-server
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
arrow==1.3.0
    # via isoduration
asttokens==2.4.1
    # via stack-data
async-lru==2.0.4
    # via jupyterlab
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
babel==2.14.0
    # via jupyterlab-server
beautifulsoup4==4.12.3
    # via nbconvert
bleach==6.1.0
    # via
    #   nbconvert
    #   panel
blosc2==2.5.1
    # via tables
bokeh==3.3.4
    # via
    #   hvplot
    #   panel
cdflib==1.2.6
    # via viresclient
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
    #   netcdf4
    #   requests
cffi==1.16.0
    # via argon2-cffi-bindings
cftime==1.6.3
    # via netcdf4
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via swarmpal
colorcet==3.1.0
    # via
    #   holoviews
    #   hvplot
comm==0.2.2
    # via ipykernel
contourpy==1.2.0
    # via
    #   bokeh
    #   matplotlib
cycler==0.12.1
    # via matplotlib
debugpy==1.8.1
    # via ipykernel
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via nbconvert
executing==2.0.1
    # via stack-data
fastjsonschema==2.19.1
    # via nbformat
fonttools==4.50.0
    # via matplotlib
fqdn==1.5.1
    # via jsonschema
h11==0.14.0
    # via httpcore
hapiclient==0.2.5
    # via swarmpal
holoviews==1.18.3
    # via hvplot
httpcore==1.0.4
    # via httpx
httpx==0.27.0
    # via jupyterlab
hvplot==0.9.2
idna==3.6
    # via
    #   anyio
    #   httpx
    #   jsonschema
    #   requests
ipykernel==6.29.3
    # via jupyterlab
ipython==8.22.2
    # via ipykernel
isodate==0.6.1
    # via hapiclient
isoduration==20.11.0
    # via jsonschema
jedi==0.19.1
    # via ipython
jinja2==3.1.3
    # via
    #   bokeh
    #   jupyter-server
    #   jupyterlab
    #   jupyterlab-server
    #   nbconvert
    #   viresclient
joblib==1.3.2
    # via hapiclient
json5==0.9.24
    # via jupyterlab-server
jsonpointer==2.4
    # via jsonschema
jsonschema==4.21.1
    # via
    #   jupyter-events
    #   jupyterlab-server
    #   nbformat
jsonschema-specifications==2023.12.1
    # via jsonschema
jupyter-client==8.6.1
    # via
    #   ipykernel
    #   jupyter-server
    #   nbclient
jupyter-core==5.7.2
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-server
    #   jupyterlab
    #   nbclient
    #   nbconvert
    #   nbformat
jupyter-events==0.9.1
    # via jupyter-server
jupyter-lsp==2.2.4
    # via jupyterlab
jupyter-server==2.13.0
    # via
    #   jupyter-lsp
    #   jupyterlab
    #   jupyterlab-server
    #   notebook-shim
jupyter-server-terminals==0.5.3
    # via jupyter-server
jupyterlab==4.1.5
jupyterlab-pygments==0.3.0
    # via nbconvert
jupyterlab-server==2.25.4
    # via jupyterlab
kiwisolver==1.4.5
    # via matplotlib
linkify-it-py==2.0.3
    # via panel
markdown==3.6
    # via panel
markdown-it-py==3.0.0
    # via
    #   mdit-py-plugins
    #   panel
markupsafe==2.1.5
    # via
    #   jinja2
    #   nbconvert
matplotlib==3.8.3
    # via swarmpal
matplotlib-inline==0.1.6
    # via
    #   ipykernel
    #   ipython
mdit-py-plugins==0.4.0
    # via panel
mdurl==0.1.2
    # via markdown-it-py
mistune==3.0.2
    # via nbconvert
msgpack==1.0.8
    # via blosc2
nbclient==0.10.0
    # via nbconvert
nbconvert==7.16.2
    # via jupyter-server
nbformat==5.10.3
    # via
    #   jupyter-server
    #   nbclient
    #   nbconvert
ndindex==1.8
    # via blosc2
nest-asyncio==1.6.0
    # via ipykernel
netcdf4==1.6.5
    # via viresclient
notebook-shim==0.2.4
    # via jupyterlab
numexpr==2.9.0
    # via tables
numpy==1.26.4
    # via
    #   blosc2
    #   bokeh
    #   cdflib
    #   cftime
    #   contourpy
    #   hapiclient
    #   holoviews
    #   hvplot
    #   matplotlib
    #   netcdf4
    #   numexpr
    #   pandas
    #   scipy
    #   swarmpal
    #   tables
    #   xarray
overrides==7.7.0
    # via jupyter-server
packaging==24.0
    # via
    #   bokeh
    #   holoviews
    #   hvplot
    #   ipykernel
    #   jupyter-server
    #   jupyterlab
    #   jupyterlab-server
    #   matplotlib
    #   nbconvert
    #   tables
    #   xarray
    #   xarray-datatree
pandas==2.2.1
    # via
    #   bokeh
    #   hapiclient
    #   holoviews
    #   hvplot
    #   panel
    #   viresclient
    #   xarray
pandocfilters==1.5.1
    # via nbconvert
panel==1.3.8
    # via
    #   holoviews
    #   hvplot
param==2.0.2
    # via
    #   holoviews
    #   hvplot
    #   panel
    #   pyviz-comms
parso==0.8.3
    # via jedi
pexpect==4.9.0
    # via ipython
pillow==10.2.0
    # via
    #   bokeh
    #   matplotlib
platformdirs==4.2.0
    # via jupyter-core
prometheus-client==0.20.0
    # via jupyter-server
prompt-toolkit==3.0.43
    # via ipython
psutil==5.9.8
    # via ipykernel
ptyprocess==0.7.0
    # via
    #   pexpect
    #   terminado
pure-eval==0.2.2
    # via stack-data
py-cpuinfo==9.0.0
    # via
    #   blosc2
    #   tables
pycparser==2.21
    # via cffi
pygments==2.17.2
    # via
    #   ipython
    #   nbconvert
pyparsing==3.1.2
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   arrow
    #   jupyter-client
    #   matplotlib
    #   pandas
python-dotenv==1.0.1
python-json-logger==2.0.7
    # via jupyter-events
pytz==2024.1
    # via pandas
pyviz-comms==3.0.1
    # via
    #   holoviews
    #   panel
pyyaml==6.0.1
    # via
    #   bokeh
    #   jupyter-events
pyzmq==25.1.2
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-server
referencing==0.34.0
    # via
    #   jsonschema
    #   jsonschema-specifications
    #   jupyter-events
requests==2.31.0
    # via
    #   jupyterlab-server
    #   panel
    #   viresclient
rfc3339-validator==0.1.4
    # via
    #   jsonschema
    #   jupyter-events
rfc3986-validator==0.1.1
    # via
    #   jsonschema
    #   jupyter-events
rpds-py==0.18.0
    # via
    #   jsonschema
    #   referencing
scipy==1.12.0
    # via swarmpal
send2trash==1.8.2
    # via jupyter-server
six==1.16.0
    # via
    #   asttokens
    #   bleach
    #   isodate
    #   python-dateutil
    #   rfc3339-validator
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
soupsieve==2.5
    # via beautifulsoup4
stack-data==0.6.3
    # via ipython
swarmpal @ git+https://github.com/Swarm-DISC/SwarmPAL.git@39e52980f36ea8290a5649a7cca5b08d7b19d10d
tables==3.9.2
    # via viresclient
terminado==0.18.1
    # via
    #   jupyter-server
    #   jupyter-server-terminals
tinycss2==1.2.1
    # via nbconvert
tornado==6.4
    # via
    #   bokeh
    #   ipykernel
    #   jupyter-client
    #   jupyter-server
    #   jupyterlab
    #   terminado
tqdm==4.66.2
    # via
    #   panel
    #   viresclient
traitlets==5.14.2
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   jupyter-client
    #   jupyter-core
    #   jupyter-events
    #   jupyter-server
    #   jupyterlab
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
types-python-dateutil==2.9.0.20240316
    # via arrow
typing-extensions==4.10.0
    # via panel
tzdata==2024.1
    # via pandas
uc-micro-py==1.0.3
    # via linkify-it-py
uri-template==1.3.0
    # via jsonschema
urllib3==2.2.1
    # via
    #   hapiclient
    #   requests
viresclient==0.11.6
    # via swarmpal
wcwidth==0.2.13
    # via prompt-toolkit
webcolors==1.13
    # via jsonschema
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
websocket-client==1.7.0
    # via jupyter-server
xarray==2024.2.0
    # via
    #   swarmpal
    #   viresclient
    #   xarray-datatree
xarray-datatree==0.0.14
    # via swarmpal
xyzservices==2023.10.1
    # via
    #   bokeh
    #   panel
